+++ packages/python/m/cli/engine/sys.py
@@ -22,22 +22,21 @@ def _get_meta_module(name: str) -> MetaModule:
     )
 
 
-def get_command_modules(root: str, commands_module: str) -> CmdMap:
+def get_command_modules(commands_module: str) -> CmdMap:
     """Create a dictionary of a file name to its module.
 
     These modules should define a `run` function.
 
     Args:
-        root: Absolute path of the directory containing the __main__.py module
-            that activates the cli.
         commands_module: The full module resolution. For instance
             `m.cli.commands`.
 
     Returns:
         A dictionary of command modules.
     """
-    dir_name = '/'.join(commands_module.split('.')[1:])
-    mod_names = list(iglob(f'{root}/{dir_name}/*.py'))
+    cmd_module = import_module(commands_module)
+    dir_name = pth.split(cmd_module.__file__ or '')[0]
+    mod_names = list(iglob(f'{dir_name}/*.py'))
     mod = {}
     for mod_name in mod_names:
         name = pth.split(mod_name)[1][:-3]
@@ -47,34 +46,29 @@ def get_command_modules(root: str, commands_module: str) -> CmdMap:
     return mod
 
 
-def get_cli_command_modules(file_path: str) -> Tuple[NestedCmdMap, MetaMap]:
+def get_cli_command_modules(commands_module: str) -> Tuple[NestedCmdMap, MetaMap]:
     """Return a dictionary containing the commands and subcommands for the cli.
 
-    `file_path` is expected to be the absolute path to the __main__.py file.
-    Another restriction is that the __main__.py file must have the
-    `cli.commands` module as its sibling.
-
     Args:
-        file_path: Absolute path to the `__main__.py` file.
+        commands_module: module containing all the commands.
 
     Returns:
         A tuple with a map containing all the modules that have the commands
         for the cli and a map with the meta data related to each command.
     """
-    root = pth.split(pth.abspath(file_path))[0]
-    main_mod = pth.split(root)[1]
-    cli_root = f'{main_mod}.cli'
-    root_cmd = get_command_modules(root, f'{cli_root}.commands')
+    cmd_module = import_module(commands_module)
+    root_cmd = get_command_modules(commands_module)
     mod: NestedCmdMap = {}
     meta: MetaMap = {}
     for key, cmd_mod in root_cmd.items():
         mod[key] = cmd_mod
-    meta['_root'] = _get_meta_module(f'{cli_root}.commands')
-    subcommands = list(iglob(f'{root}/cli/commands/*'))
+    meta['_root'] = _get_meta_module(commands_module)
+    root = pth.split(cmd_module.__file__ or '')[0]
+    subcommands = list(iglob(f'{root}/*'))
     for cmd_name in subcommands:
         if cmd_name.endswith('.py') or cmd_name.endswith('__'):
             continue
         name = pth.split(cmd_name)[1]
-        mod[name] = get_command_modules(root, f'{cli_root}.commands.{name}')
-        meta[name] = _get_meta_module(f'{cli_root}.commands.{name}')
+        mod[name] = get_command_modules(f'{commands_module}.{name}')
+        meta[name] = _get_meta_module(f'{commands_module}.{name}')
     return mod, meta